Migrate from org config to elisp, use-package to Cask
For a variety of reasons, I've decided to migrate away from using an org file to
configure Emacs. For one, I found that I would just get lost in writing about
the logic for my configurations, without actually finding any value in doing so.
I also found that whenever I changed my config, it would become a hastle to
update the text that describes what the code is doing, let alone the code
itself. I've now switched to a multi-file based approach, with lots of
descriptively named functions.

I also switched from use-package to Cask for managing dependencies. I find it
very useful to be able to just have a list of all the dependencies, and be able
to configure them independently and have them all available in my config files.